+++ sphinx/application/data/features/feature-repository/src/main/java/chat/sphinx/feature_repository/mappers/message/MessageDboPresenterMapper.kt
@@ -22,10 +22,10 @@ internal class MessageDboPresenterMapper(
                         //Old Podcast boost with message type and text format
                         decrypted.isPodBoost -> {
                             withContext(default) {
-                                decrypted.value.replaceFirst(PodBoost.MESSAGE_PREFIX, "")
+                                decrypted.value.replaceFirst(FeedBoost.MESSAGE_PREFIX, "")
                                     .toPodBoostOrNull(moshi)
                                     ?.let { podBoost ->
-                                        message._podBoost = podBoost
+                                        message._feedBoost = podBoost
                                     }
                             }
                         }
@@ -46,9 +46,9 @@ internal class MessageDboPresenterMapper(
                 } else if (message.type.isBoost() && message.replyUUID == null) {
                     //New Podcast boost with boost type (29) and null uuid
                     withContext(default) {
-                        decrypted.value.replaceFirst(PodBoost.MESSAGE_PREFIX, "")
+                        decrypted.value.replaceFirst(FeedBoost.MESSAGE_PREFIX, "")
                             .toPodBoostOrNull(moshi)?.let { podBoost ->
-                                message._podBoost = podBoost
+                                message._feedBoost = podBoost
                             }
                     }
                 }